+++ rm_gimbal_controllers/CHANGELOG.rst
@@ -2,6 +2,47 @@
 Changelog for package rm_gimbal_controllers
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.1.11 (2023-06-20)
+-------------------
+* Merge pull request `#135 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/135>`_ from ye-luo-xi-tui/feedforward
+  Add input feedforward and fix a bug in computing desire vel at TRACK mode
+* Add input feedforward and fix a bug in computing desire vel at TRACK mode.
+* Merge pull request `#124 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/124>`_ from ye-luo-xi-tui/resistance_compensation
+  Add velocity_dead_zone and effort_dead_zone
+* Merge branch 'master' into dev/balance
+* Add velocity_dead_zone and effort_dead_zone.
+* Merge pull request `#120 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/120>`_ from ye-luo-xi-tui/master
+  0.1.10
+* Merge branch 'rm-controls:master' into gpio_calibration_controller
+* Contributors: 1moule, ye-luo-xi-tui, yezi, yuchen
+
+0.1.10 (2023-03-25)
+-------------------
+* Merge branch 'rm-controls:master' into master
+* Merge pull request `#114 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/114>`_ from ye-luo-xi-tui/resistance_compensation
+  Add resistance compensation on yaw
+* Add resistance compensation on yaw.
+* Merge pull request `#113 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/113>`_ from ye-luo-xi-tui/master
+  Use Vector3WithFilter in rm_common instead
+* Use Vector3WithFilter in rm_common instead.
+* Merge pull request `#106 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/106>`_ from ye-luo-xi-tui/master
+  0.1.9
+* Contributors: ye-luo-xi-tui, yezi
+
+0.1.9 (2023-02-21)
+------------------
+* Merge pull request `#104 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/104>`_ from ye-luo-xi-tui/balance_standard
+  Fix bug.
+* Fix bug.
+* Merge pull request `#102 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/102>`_ from ljq-lv/fix_warning
+  Fixed the bug of gimbal warning
+* Add the else to judge mode TRACK
+* Fixed the bug of gimbal warning
+* Merge branch 'master' into balance_standard
+* Merge pull request `#97 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/97>`_ from ye-luo-xi-tui/master
+  0.1.8
+* Contributors: ljq-lv, ye-luo-xi-tui, yezi
+
 0.1.8 (2022-11-24)
 ------------------
 * Merge pull request `#92 <https://github.com/ye-luo-xi-tui/rm_controllers/issues/92>`_ from ye-luo-xi-tui/target_velocity_correction